All necessary files for what I have worked on so far,
You're going to need to add the csv file yourself to the release and debug directory since it's too large. Also you're going to need to get sfml and configure it for the file. If you have visual studio I can show you how to do it because it's actually quite annoying to set up first tine.